+++ src/resolve-environment-action.ts
@@ -7,7 +7,7 @@ import {
   getTemporaryDirectory,
 } from "./actions-util";
 import { getGitHubVersion } from "./api-client";
-import { CommandInvocationError } from "./codeql";
+import { CommandInvocationError } from "./cli-errors";
 import * as configUtils from "./config-utils";
 import { getActionsLogger } from "./logging";
 import { runResolveBuildEnvironment } from "./resolve-environment";
@@ -32,18 +32,14 @@ async function run() {
   const logger = getActionsLogger();
 
   try {
-    if (
-      !(await sendStatusReport(
-        await createStatusReportBase(
-          ACTION_NAME,
-          "starting",
-          startedAt,
-          await checkDiskUsage(logger),
-        ),
-      ))
-    ) {
-      return;
-    }
+    await sendStatusReport(
+      await createStatusReportBase(
+        ACTION_NAME,
+        "starting",
+        startedAt,
+        await checkDiskUsage(logger),
+      ),
+    );
 
     const gitHubVersion = await getGitHubVersion();
     checkGitHubVersionInRange(gitHubVersion, logger);