+++ src/upload-lib.ts
@@ -14,7 +14,12 @@ import * as fingerprints from "./fingerprints";
 import { Logger } from "./logging";
 import { parseRepositoryNwo, RepositoryNwo } from "./repository";
 import * as util from "./util";
-import { SarifFile, UserError, wrapError } from "./util";
+import { SarifFile, ConfigurationError, wrapError } from "./util";
+
+const GENERIC_403_MSG =
+  "The repo on which this action is running has not opted-in to CodeQL code scanning.";
+const GENERIC_404_MSG =
+  "The CodeQL code scanning feature is forbidden on this repository.";
 
 // Takes a list of paths to sarif files and combines them together,
 // returning the contents of the combined sarif file.
@@ -107,19 +112,35 @@ async function uploadPayload(
 
   const client = api.getApiClient();
 
-  const response = await client.request(
-    "PUT /repos/:owner/:repo/code-scanning/analysis",
-    {
-      owner: repositoryNwo.owner,
-      repo: repositoryNwo.repo,
-      data: payload,
-    },
-  );
+  try {
+    const response = await client.request(
+      "PUT /repos/:owner/:repo/code-scanning/analysis",
+      {
+        owner: repositoryNwo.owner,
+        repo: repositoryNwo.repo,
+        data: payload,
+      },
+    );
 
-  logger.debug(`response status: ${response.status}`);
-  logger.info("Successfully uploaded results");
-
-  return response.data.id;
+    logger.debug(`response status: ${response.status}`);
+    logger.info("Successfully uploaded results");
+    return response.data.id;
+  } catch (e) {
+    if (util.isHTTPError(e)) {
+      switch (e.status) {
+        case 403:
+          core.warning(e.message || GENERIC_403_MSG);
+          break;
+        case 404:
+          core.warning(e.message || GENERIC_404_MSG);
+          break;
+        default:
+          core.warning(e.message);
+          break;
+      }
+    }
+    throw e;
+  }
 }
 
 export interface UploadStatusReport {
@@ -158,7 +179,7 @@ export function findSarifFilesInDir(sarifPath: string): string[] {
  * Uploads a single SARIF file or a directory of SARIF files depending on what `sarifPath` refers
  * to.
  *
- * @param considerInvalidRequestUserError Whether an invalid request, for example one with a
+ * @param considerInvalidRequestConfigError Whether an invalid request, for example one with a
  *                                        `sarifPath` that does not exist, should be considered a
  *                                        user error.
  */
@@ -168,8 +189,8 @@ export async function uploadFromActions(
   category: string | undefined,
   logger: Logger,
   {
-    considerInvalidRequestUserError,
-  }: { considerInvalidRequestUserError: boolean },
+    considerInvalidRequestConfigError: considerInvalidRequestConfigError,
+  }: { considerInvalidRequestConfigError: boolean },
 ): Promise<UploadResult> {
   try {
     return await uploadFiles(
@@ -187,8 +208,8 @@ export async function uploadFromActions(
       logger,
     );
   } catch (e) {
-    if (e instanceof InvalidRequestError && considerInvalidRequestUserError) {
-      throw new UserError(e.message);
+    if (e instanceof InvalidRequestError && considerInvalidRequestConfigError) {
+      throw new ConfigurationError(e.message);
     }
     throw e;
   }
@@ -489,8 +510,8 @@ export async function waitForProcessing(
         break;
       } else if (status === "failed") {
         const message = `Code Scanning could not process the submitted SARIF file:\n${response.data.errors}`;
-        throw shouldConsiderAsUserError(response.data.errors as string[])
-          ? new UserError(message)
+        throw shouldConsiderConfigurationError(response.data.errors as string[])
+          ? new ConfigurationError(message)
           : new InvalidRequestError(message);
       } else {
         util.assertNever(status);
@@ -508,7 +529,7 @@ export async function waitForProcessing(
 /**
  * Returns whether the provided processing errors should be considered a user error.
  */
-function shouldConsiderAsUserError(processingErrors: string[]): boolean {
+function shouldConsiderConfigurationError(processingErrors: string[]): boolean {
   return (
     processingErrors.length === 1 &&
     processingErrors[0] ===